+++ src/fhirpath/utils.py
@@ -24,15 +24,13 @@ from zope.interface import implementer
 from fhirpath.thirdparty import Proxy
 
 from .enums import FHIR_VERSION
-from .interfaces import IModel
 from .interfaces import IPathInfoContext
-from .navigator import PathNavigator
 from .storage import FHIR_RESOURCE_CLASS_STORAGE
 from .storage import PATH_INFO_STORAGE
 from .types import PrimitiveDataTypes
 
 
-__author__ = "Md Nazrul Islam<email2nazrul@gmail.com>"
+__author__ = "Md Nazrul Islam <email2nazrul@gmail.com>"
 
 
 def _reraise(tp, value, tb=None):
@@ -442,46 +440,6 @@ class PathInfoContextProxy(Proxy):
         """ """
         super(PathInfoContextProxy, self).__init__()
         self.initialize(context)
-
-
-class ModelFactory(type):
-    """FHIR Model factory"""
-
-    def __new__(cls, name, bases, attrs, **kwargs):
-        super_new = super().__new__
-
-        # xxx: customize module path?
-        module = attrs.pop("__module__", cls.__module__)
-        new_attrs = {"__module__": module}
-        classcell = attrs.pop("__classcell__", None)
-        if classcell is not None:
-            new_attrs["__classcell__"] = classcell
-
-        new_class = super_new(cls, name, bases, new_attrs, **kwargs)
-
-        # Attach Interface
-        new_class = implementer(IModel)(new_class)
-
-        return new_class
-
-    def add_to_class(cls, name, value):
-        """ """
-        setattr(cls, name, value)
-
-
-class Model:
-    """ """
-
-    @staticmethod
-    def create(resource_type: Text, fhir_version: FHIR_VERSION = FHIR_VERSION.DEFAULT):
-        """ """
-        klass = import_string(
-            cast(Text, lookup_fhir_class_path(resource_type, fhir_release=fhir_version))
-        )
-        # xxx: should be cache?
-        model = ModelFactory(f"{klass.__name__}Model", (klass, PathNavigator), {})
-
-        return model
 
 
 class BundleWrapper: