+++ HISTORY.rst
@@ -2,7 +2,7 @@
 History
 =======
 
-0.5.0 (unreleased)
+0.5.0 (2020-03-11)
 ------------------
 
 New Features